requirements.txt
scipy
liac-arff
sklearn
scikit-multilearn